+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ next
 #. Pin django-ckeditor to >= 3.6.2
 #. Remember me field now on login and join forms. Checked by default.
 #. Any call to get_XXX_url is now layer aware.
+#. Comment posting now ajaxified depending on browser capabilities.
 
 0.2.1
 -----